+++ example/server.ts
@@ -1,11 +1,15 @@
 import { Elysia, t, type SCHEMA } from 'elysia'
+import { cors } from '@elysiajs/cors'
 import { websocket } from '@elysiajs/websocket'
 
 const app = new Elysia()
+    .use(cors())
     .use(websocket())
     .get('/', () => 'Elysia')
     .post('/', () => 'Elysia')
     .get('/sign-in', () => 'ok')
+    .get('/a/a/:b/:c/d', () => 'Ok')
+    .get('/b/:c', () => 'Ok')
     .post('/products/nendoroid', ({ body }) => body, {
         schema: {
             body: t.Object({
@@ -35,6 +39,15 @@ const app = new Elysia()
         },
         message(ws, message) {
             ws.send(message)
+        }
+    })
+    .ws('/chat/:room/:name', {
+        message(ws, message) {
+            ws.send(message)
+        },
+        schema: {
+            body: t.String(),
+            response: t.String()
         }
     })
     .listen(8080)